Created the stepVector(double x, double y) function. Essentially, it sees to it that the given vector (passed in terms of components) has a magnitude of 0.06 (the step size specified in Simulator.hpp), preserving direction.
Modified the Bug0 algorithm accordingly to ensure that any vectors
returned to the simulator have magnitude 0.06, which necessitated
changing a few of the other functions to avoid mathematical
redundencies.

Clarified a few comments and updated for accuracy, mostly with regard to
demonstration of specific vector formulae.

After a few tests run on all the obstacle files, the bug works a lot better.
It doesn't regularly phase through walls, at least. It got a bit confused
on the spiral map (obstacles4.txt), since it can't figure out if an obstacle's
in between it and the goal yet.

Nothing else worthy of a message to report.

Hah, made you look.